+++ packages/invoice-adapter-amego/CHANGELOG.md
@@ -119,6 +119,6 @@ Co-Authored-By: Claude <noreply@anthropic.com>
 - **invoice-adapter-amego:** implementation of invalidAllowance ([8f34e00](https://github.com/Rytass/Utils/commit/8f34e004ef657a35f0e0702315c77a3d62fc3a95))
 - **invoice-adapter-amego:** refactor testing code to using mock data ([b6febfa](https://github.com/Rytass/Utils/commit/b6febfa60e0aa17cf008a5de2b55706d66bf63d2))
 - **invoice-adapter-amego:** update invoice package version ([4f86340](https://github.com/Rytass/Utils/commit/4f86340f19cb13b76a65353ecb63fb70958eaaab))
-- **invoice-adapter-amego:** use verifyVatNumber function to check vatNumber ([074fa47](https://github.com/Rytass/Utils/commit/074fa47af73a16cde6f205b23a3898e5acb16539))
+- **invoice-adapter-amego:** use isValidVATNumber function to check vatNumber ([074fa47](https://github.com/Rytass/Utils/commit/074fa47af73a16cde6f205b23a3898e5acb16539))
 - **invoice:** implementation of invoice provider amego, isMobileBarcodeValid ([5509876](https://github.com/Rytass/Utils/commit/550987604eb3e477eba3ac74cb95b0b534b6321b))
 - **invoice:** implementation of invoice provider amego, issue, void ([3601090](https://github.com/Rytass/Utils/commit/3601090b1ed0fae170f3e8598e972619577c7819))